+++ src/views/team-info.js
@@ -3,7 +3,7 @@ import GraphSet from '../components/graph-set'
 import TeamInfoParser from '../data-parsing/team-info-parser'
 import TeamInfoConfig from '../data-parsing/configs/team-info-config'
 import MainMenu from './main-menu'
-import {pickScheme} from '../components/charts/chart-builder'
+import {pickScheme} from '../components/charts/chart-utils'
 import axios from 'axios'
 
 class TeamInfo extends Component {
